+++ src/BankAccountNormalizedInterface.php
@@ -4,6 +4,8 @@ namespace Cs278\BankModulus;
 
 interface BankAccountNormalizedInterface extends BankAccountInterface
 {
+    const ACCOUNT_NUMBER_LENGTH = 8;
+
     /** @return BankAccountInterface */
     public function getOriginalBankAccount();
 }